Extracted template from site-menu component
