+++ src/components/site-menu/index.js
@@ -1,0 +1,8 @@
+import Component from "../component";
+import {template} from "./template";
+
+export default class SiteMenu extends Component {
+  constructor() {
+    super(template());
+  }
+}
